refactor: interface OCP 적용
