+++ src/main/java/chess/dao/Dao.java
@@ -1,0 +1,6 @@
+package chess.dao;
+
+public interface Dao<T> {
+
+    T save(T target);
+}
